requirements.txt
MetaTrader5>=5.0.45
openai>=1.40.0
streamlit>=1.35.0
pandas>=2.2.2
numpy>=1.26.4
python-dotenv>=1.0.1
pydantic>=2.8.2
